+++ src/main/java/eu/sia/demo/mem/usage/core/PerformanceMetricExtractor.java
@@ -5,5 +5,5 @@ import java.util.concurrent.TimeUnit;
 
 public interface PerformanceMetricExtractor {
 	PerformanceStatistic calculate(String name, List<MetricEntry> list);
-	PerformanceStatistic calculate(String name, List<MetricEntry> list, int timeDelta, TimeUnit timeUnit);
+	PerformanceStatistic calculate(String name, List<MetricEntry> list, long timeDelta, TimeUnit timeUnit);
 }
