+++ src/main/java/eu/sia/demo/mem/usage/util/TimeUtil.java
@@ -5,7 +5,7 @@ import java.util.concurrent.TimeUnit;
 import java.util.function.Function;
 
 public interface TimeUtil {
-	long getLowest(int timeDelta, TimeUnit timeunit);
+	long getLowest(long timeDelta, TimeUnit timeunit);
 	Float nanoToMilli(Long nanoDelta);
 	String toTimeStamp(Calendar calendar);
 	Function<Calendar, String> getToTimeStampFunction();
