+++ src/main/java/eu/sia/demo/mem/usage/view/behaviour/StatisticLayoutCreator.java
@@ -5,5 +5,5 @@ import java.util.concurrent.TimeUnit;
 import eu.sia.demo.mem.usage.view.RefreshableComponent;
 
 public interface StatisticLayoutCreator {
-	RefreshableComponent create(int removeDelta, TimeUnit timeunit);
+	RefreshableComponent create(long removeDelta, TimeUnit timeunit);
 }
